app/requirements.txt
watchdog>=3.0.0
psutil>=5.9.0
pillow>=10.0.0
pystray>=0.19.4
pyinstaller>=6.0.0
pywin32>=306; sys_platform == "win32"
